pyproject.toml
[tool.poetry]
name = "valhalla"
version = "0.3.0"
description = ""
authors = ["Matthew Messinger <mattmess1221@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "^0.8.0"
Authlib = "^1.0.1"
expiringdict = "^1.2.0"
fastapi = "^0.79.0"
fs = "^2.4.11"
fs-s3fs = "^1.1.1"
gunicorn = "^20.1.0"
httpx = "^0.23.0"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.2"
passlib = "^1.7.2"
Pillow = "^9.2.0"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
raygun4py = "^4.3.0"
SQLAlchemy = "^1.4.39"
SQLAlchemy-Utils = "^0.38.3"
uvicorn = "^0.18.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
